+++ firebase_core/firebase_core_vm/lib/src/platform_js.dart
@@ -1,0 +1,7 @@
+// File created by
+// Lung Razvan <long1eu>
+// on 04/03/2020
+
+const bool isMobile = false;
+const bool isDesktop = false;
+const bool isWeb = true;