+++ firebase_core/firebase_core_vm/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.4
+
+- Expose `isWeb`, `isMobile` and `isDesktop` based on conditional imports
+- Throw unimplemented error when trying to create a FirebaseApp in a js/html context
+
 ## 0.0.3
 
 - Rename library from `firebase_common` to `firebase_core`